src/requirements.txt
appdirs==1.4.4
boto3==1.26.85
botocore==1.29.85
certifi==2022.12.7
charset-normalizer==3.0.1
click==7.1.2
contextlib2==21.6.0
curator-opensearch==0.0.9
idna==3.4
jmespath==1.0.1
more-itertools==8.13.0
opensearch-py==2.2.0
ordered-set==4.0.2
packaging==21.3
pep517==0.12.0
pydash==6.0.2
pyparsing==2.4.7
python-dateutil==2.8.2
PyYAML==5.4.1
requests==2.28.2
requests-aws4auth==1.2.2
retrying==1.3.3
s3transfer==0.6.0
six==1.16.0
tomli==2.0.1
urllib3==1.26.14
voluptuous==0.13.1
